Refactor shared visual sub-components into their own folderr
