+++ src/components/shared/AutoSuggestText.tsx
@@ -17,11 +17,11 @@ import TextInput from 'react-autocomplete-input';
 import 'react-autocomplete-input/dist/bundle.css';
 
 // Utils:
-import { ParamPrefixes, ParamDefaults, GameFilterParams } from '../utils/FilterModels';
-import { RosterCompareModel } from '../components/RosterCompareTable';
-import { dataLastUpdated } from '../utils/internal-data/dataLastUpdated';
-import { ClientRequestCache } from '../utils/ClientRequestCache';
-import { QueryUtils } from "../utils/QueryUtils";
+import { ParamPrefixes, ParamDefaults, GameFilterParams } from '../../utils/FilterModels';
+import { RosterCompareModel } from '../../components/RosterCompareTable';
+import { dataLastUpdated } from '../../utils/internal-data/dataLastUpdated';
+import { ClientRequestCache } from '../../utils/ClientRequestCache';
+import { QueryUtils } from "../../utils/QueryUtils";
 
 /** The keydown event does not come from AutoSuggestText element */
 export const notFromAutoSuggest = (event: any) => {
