+++ src/components/shared/GenericCollapsibleCard.tsx
