+++ src/components/shared/__tests__/HistorySelector.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import HistorySelector from '../HistorySelector';
 import { shallow } from 'enzyme'
 import toJson from 'enzyme-to-json'
-import { HistoryManager } from '../../utils/HistoryManager';
+import { HistoryManager } from '../../../utils/HistoryManager';
 
 describe("HistorySelector", () => {
   test("HistorySelector - should create snapshot", () => {
